requirements.txt
pandas
openpyxl
python-dotenv
xlsxwriter
kaleido
plotly